consolidate csrf logic into single file
