+++ csrf.go
@@ -3,9 +3,26 @@ package goa4web
 import (
 	"crypto/sha256"
 	"net/url"
+	"os"
+	"strings"
 
+	config "github.com/arran4/goa4web/config"
 	"github.com/gorilla/csrf"
 )
+
+// csrfEnabled reports if CSRF protection should be active.
+func csrfEnabled() bool {
+	v := strings.ToLower(os.Getenv(config.EnvCSRFEnabled))
+	if v == "" {
+		return true
+	}
+	switch v {
+	case "0", "false", "off", "no":
+		return false
+	default:
+		return true
+	}
+}
 
 // newCSRFMiddleware returns a wrapper enforcing CSRF protection using the
 // provided session secret and HTTP configuration.
